Allow to check whether an entity action is enabled or not
Previously, checking for entity action enablement in the respective controller
method was not directly supported by this plugin.
This change adds a new component that provides a method to do exactly
that. In addition, a behavior was added providing a finder that adds
associations required by all or a specific entity action of the tables
entity to the query.